requirements.txt
python-socketio==5.11.4
Requests==2.32.3
streamlit==1.37.0
